Block timestamps
To provide more information about the time a particular block was
created, include a best-effort timestamp in the block.

This timestamp is the median of all the *valid* timestamps (timestamps
not older than the last block timestamp) inserted into the RBC payload
by each consensus member. The stamp information also includes what that
member thought this head of the chain was at the time it started running
the hbbft round. This extra information allows us to detect 'stale'
events from a previous round that we've already been gossiped a block
for.