Fix extension scripts: avoid warning on serial console, have exit status
Extension scripts are expected to call exit themselves, or the runner
script will exit with code 1. This breaks using the miner extension
commands in shell scripts. To fix this we exit with the exit code of the
RPC call (if the RPC call does not return with `ok`, the exit status
will be non-zero).