validation sidecar for hbbft groups to prevent blocking
The current transaction validation logic is serial and blocks the
process that is meant to handle network messages for honeybader, so
long validations can signifcantly delay message processing and delay
procotol round completion.

This change creates a new process that does limited width parallel
validation of slow transactions, allowing the honeybadger protocol
handler to proceed (mostly) asynchronously.  In addition it fixes a
bug with the consensus manager that would allow a group that should be
stopped (on a node leaving the consensus group) to keep running.