Clean up test data and logs on common test success
Instead of leaving behind dozens of gigabytes of logs and
blockchain/ledger databases for each test run, if the test suceeds.
delete all the logs and databases at the end of the run.

Additionally use init/end per testcase functions for the poc suite so
that it better matches the other suites and ensures node shutdown after
the test finishes.