Add LoRaWAN support
This patchset adds support for routing LoRaWAN packets.

The specific mechanism used to make this work is as follows:

For LoRaWAN join packets, the APPEUI consists of
<<OUI:32/integer-unsigned-big, DeviceID:32/integer-unsigned-big>>, this
is equivalent to the LongFi address field.

On join, the router returns the device a DevAddr, which is a 32 bit
identifier. Unfortunately with this constrained address space, we have
to convert out 64 bit address space into a 32 bit one. This is not ideal
but we can use a trick where we give all devices the same DevAddr and
figure out which device it is on the router side by checking the message
integrity code against all known session keys for that OUI. This takes
about 32 microseconds per key, so it's not terrible.

Thus, join packets are routed by the top half of the AppEUI and other
packets are routed by the DevAddr.

As part of this work we have also switched from using concentrate to
using the stock Semtech packet forwarder and its protocol:
https://github.com/Lora-net/packet_forwarder/blob/master/PROTOCOL.TXT

Finally, we have also changed the channel scheme to use LoRaWAN US 915
channel group 7.