Make election_done asynchronous to avoid mutual blocking consensus mgr
Instead of doing a gen_server:call and blocking on the reply, do a
gen_server cast and wait for a `mark_done` event to come back so we know
we've handed the key off. To make this work without sending duplicate
election_done messages to the consensus manager we track whether the
election done was acknowleded with a seperate record field.

This PR also attempts to improve the logging slightly to include
election delay adjacent to the election height more consistently and it
also tries to improve the group cancel behaviour.

Finally this bumps core for a chain var default value fix.

B