Add CLI support for querying miner manufacturing info.
With the adoption of HIP-19 there are now several manufacturers of approved
Helium miners, and thus, no single central staking party to approve onboarding
transactions, such as "assert_gateway" and "assert_location". The new process
requires gateway owners to find the appropriate staking payer for their
particular gateway when constructing such transactions.

A new API server has been created to manage this process, and to use it,
users must submit an "onboarding" key to the server to learn the ultimate
manufacturer and responsible staking party for the gateway. This code
adds support for querying that API directly from the miner in order to
show the user the information they will need for such transactions.