Fix JSON <-> Erlang string issues.
Fix several points in the JSON/RPC interface where items which
should be strings in JSON are instead appearing as lists of integers
in result sets.

Likewise, fix an issue in the opposite direction, for `get_account`,
where the account address wasn't being B58 decoded before being
handed off to the ledger for querying.