Fix validator name so that it is returned as a string, add address.
Fix the `ledger_validators` method so that the return type of the "name"
attribute for a validator appears properly in the JSON results as a
string instead of a list of integers.

Also add the validator's address to the result set, since that is
an important identifier to have.

Fixes helium/miner#908.