Update ledger_validators, make more JSON semantic, adjust "verbose".
Update the `ledger_validators` call to:
* Return more JSON-native datatypes, especially integers and floats,
  where possible.
* Always include an `address` element
* Optionally include a `name` element and `nonce` element if
  the `verbose` parameter is set.

This output is slightly different than the previous version, which
would return the validator address in `validator_address`, and even
then, only if `verbose` had been set.

I believe that `address` is a critical identity property to the
validator, not an optional one. Human-readable angry-purple-tiger
name, on the other hand, is optional instead.