Refactor hbbft_perf, renaming some confusing variables.
Refactor and clean up miner_util:hbbft_perf/0 so that it returns
a map, rather than a giant tuple, and rename some of its internal
variables so that they are easier to understand (e.g. "Start0" is
now "ElectionStart").

Also clean up the JSON/RPC and the CLI `hbbft perf` endpoints as
well.