Remove link to dead docs.
As far as I can tell, the 'run a miner' content in docs is gone. I'd be happy to be proven wrong.

Proposing removing the whole block instead of pointing to the gateway-rs guides. It could be nice to reference quay builds in this doc, but it can balloon into a bigger discussion around packet forwarders, etc.

Also took the opportunity to do a little formatting cleanup.