Fix a bug in denylist checking that can cause unintended rollbacks.
Denylist freshness relies on comparing the latest available version
with the version the miner currently has. This comparison uses
github tags for versioning, but due to a bug in JSON parsing, any
new version with a _different_ tag than what the miner currently has,
even if it appears to be an _older_ tag, will be accepted as newer,
and will cause the miner to replace the current denylist with the
older version, even if that was unintended.