Begin centralized sys.config for Docker-based testnet miner.
After understanding a bit more how relx, rebar, and hotspot
vendors collide at runtime when using containerized (Docker-based)
miners, this change begins the process whereby we will make sure that
there is only one sys.config file shipped in any containerized
product and that the runtime "overlay" feature, where additional
changes are applied to the base sys.config file at runtime, is
reserved entirely for changes that others wish to make to the
container image.

To begin this change, I am simply starting with a copy of the
main sys.config file and making that copy in such a way as to preseve
its history in git.