fix rssi computation from rssic if SNR < 0
rssis is supposed to be the useful signal strength
rssic is supposed to be the received signal strength

The blockchain excepts the rssi to be the rssis. However when falling back to rssic when rssis is unavailable, the computation is wrong when SNR < 0
Indeed, in that case the useful signal strenght is below the noise level, which is rssic. So the rssis = SNR + rssic when SNR < 0
This should help use the appropriate GWMPv2 value from the packet forwarders